Added another gender option and changed maxPrice to an int

+++ src/application/Adult.java
@@ -17,7 +17,7 @@ public class Adult extends Person {
 	 * @param maxPrice the maximum price the gift sender wants to spend
 	 * @param gender the gender of the gift recipient
 	 */
-	public Adult(int age, double maxPrice, String gender) {
+	public Adult(int age, int maxPrice, String gender) {
 		super(age, maxPrice, gender);
 		interestList = new ArrayList<RatedInterest>();
 		interestList.add(new RatedInterest("Food/cooking"));
